// React/app/GerAdv_TS/AgendaRepetir/Interfaces/interface.AgendaRepetir.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
                     // copyright © 2000-2025 Menphis - Sistemas Inteligentes
                     // This file is part of the Source Genesys project                     
                    'use client';

export interface IAgendaRepetir {
// 202501251
    id: number;
 
	advogado: number,
	cliente: number,
	funcionario: number,
	processo: number,
	datafinal: string,
	horafinal: string,
	pessoal: boolean,
	frequencia: number,
	dia: number,
	mes: number,
	hora: string,
	idquem: number,
	idquem2: number,
	mensagem: string,
	idtipo: number,
	id1: number,
	id2: number,
	id3: number,
	id4: number,
	segunda: boolean,
	quarta: boolean,
	quinta: boolean,
	sexta: boolean,
	sabado: boolean,
	domingo: boolean,
	terca: boolean,
}

export interface IAgendaRepetirFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IAgendaRepetirIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}